Fix build with latest Mirage (#40)
The build rotted away, but now it's fixed.

The GPC account has been migrated to a real account, which means the images are persisting again.

Mirage is updated to 3.5.0, along with the attendant fixes. We're also using the official OCaml scripts, so no more nasty bash hacks (though they still remain for nostalgia's sake).